+++ packages/extension-base/src/background/KoniTypes.ts
@@ -40,6 +40,7 @@ import { HexString } from '@polkadot/util/types';
 
 import { EarningImpactResult } from '../services/earning-service/handlers/native-staking/dtao';
 import { GovVoteRequest, GovVotingInfo, RemoveVoteRequest, UnlockVoteRequest } from '../services/open-gov/interface';
+import { RequestAddSubstrateProxyAccount, RequestGetSubstrateProxyAccountGroup, RequestRemoveSubstrateProxyAccount, SubstrateProxyAccountGroup } from '../types/substrateProxyAccount';
 import { TransactionWarning } from './warnings/TransactionWarning';
 
 export enum RuntimeEnvironment {
@@ -584,9 +585,13 @@ export enum ExtrinsicType {
   // OPEN GOV
   GOV_VOTE = 'gov.vote',
   GOV_UNVOTE= 'gov.unvote',
-  GOV_UNLOCK_VOTE='gov.unlock-vote',
+  GOV_UNLOCK_VOTE='gov.unlock_vote',
 
   EVM_EXECUTE = 'evm.execute',
+
+  ADD_SUBSTRATE_PROXY_ACCOUNT = 'substrateProxyAccount.add',
+  REMOVE_SUBSTRATE_PROXY_ACCOUNT = 'substrateProxyAccount.remove',
+
   UNKNOWN = 'unknown'
 }
 
@@ -643,11 +648,17 @@ export interface ExtrinsicDataTypeMap {
 
   [ExtrinsicType.EVM_EXECUTE]: TransactionConfig,
   [ExtrinsicType.CROWDLOAN]: any,
+  [ExtrinsicType.SWAP]: SwapTxData,
+
   [ExtrinsicType.GOV_VOTE]: GovVoteRequest,
   [ExtrinsicType.GOV_UNVOTE]: RemoveVoteRequest,
   [ExtrinsicType.GOV_UNLOCK_VOTE]: UnlockVoteRequest,
 
+  [ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT]: RequestAddSubstrateProxyAccount,
+  [ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT]: RequestRemoveSubstrateProxyAccount,
+
   [ExtrinsicType.SWAP]: SwapTxData
+
   [ExtrinsicType.UNKNOWN]: any
 }
 
@@ -779,6 +790,7 @@ export interface TransactionHistoryItem<ET extends ExtrinsicType = ExtrinsicType
   addressPrefix?: number,
   processId?: string;
   apiTxIndex?: number;
+  substrateProxyAddresses?: string[];
 }
 
 export interface SWWarning {
@@ -1002,7 +1014,7 @@ export type TxResultType = {
   feeSymbol?: string;
 }
 
-export interface NftTransactionRequest {
+export interface NftTransactionRequest extends BaseRequestSign {
   networkKey: string,
   senderAddress: string,
   recipientAddress: string,
@@ -1085,6 +1097,7 @@ export interface SubstrateNftSubmitTransaction extends BaseRequestSign {
   senderAddress: string;
   nftItemName?: string;
   recipientAddress: string;
+  nftItem: NftItem;
 }
 
 export type RequestSubstrateNftSubmitTransaction = InternalRequestSign<SubstrateNftSubmitTransaction>;
@@ -2782,6 +2795,11 @@ export interface KoniRequestSignatures {
   'pri(openGov.unvote)': [RemoveVoteRequest, SWTransactionResponse]
   'pri(openGov.subscribeGovLockedInfo)': [null, GovVotingInfo[], GovVotingInfo[]];
   'pri(openGov.unlockVote)': [UnlockVoteRequest, SWTransactionResponse];
+
+  /* Proxy Account */
+  'pri(substrateProxyAccount.getGroupInfo)': [RequestGetSubstrateProxyAccountGroup, SubstrateProxyAccountGroup];
+  'pri(substrateProxyAccount.add)': [RequestAddSubstrateProxyAccount, SWTransactionResponse];
+  'pri(substrateProxyAccount.remove)': [RequestRemoveSubstrateProxyAccount, SWTransactionResponse];
 }
 
 export interface ApplicationMetadataType {