+++ packages/extension-base/src/services/transaction-service/index.ts
@@ -20,7 +20,7 @@ import { OptionalSWTransaction, SWDutchTransaction, SWDutchTransactionInput, SWP
 import { getExplorerLink, parseTransactionData } from '@subwallet/extension-base/services/transaction-service/utils';
 import { isWalletConnectRequest } from '@subwallet/extension-base/services/wallet-connect-service/helpers';
 import { AccountJson, BaseStepType, BasicTxErrorType, BasicTxWarningCode, BriefProcessStep, LeavePoolAdditionalData, PermitSwapData, ProcessStep, ProcessTransactionData, RequestStakePoolingBonding, RequestYieldStepSubmit, SpecialYieldPoolInfo, StepStatus, SubmitBittensorChangeValidatorStaking, SubmitJoinNominationPool, SubstrateTipInfo, TransactionErrorType, Web3Transaction, YieldPoolType } from '@subwallet/extension-base/types';
-import { anyNumberToBN, pairToAccount, reformatAddress } from '@subwallet/extension-base/utils';
+import { anyNumberToBN, isSameAddress, pairToAccount, reformatAddress } from '@subwallet/extension-base/utils';
 import { mergeTransactionAndSignature } from '@subwallet/extension-base/utils/eth/mergeTransactionAndSignature';
 import { isContractAddress, parseContractInput } from '@subwallet/extension-base/utils/eth/parseTransaction';
 import { getId } from '@subwallet/extension-base/utils/getId';
@@ -152,20 +152,30 @@ export default class TransactionService {
     const nonNativeTokenPayFeeInfo = isNonNativeTokenPayFee ? this.state.chainService.getAssetBySlug(tokenPayFeeSlug) : undefined;
     const priceMap = (await this.state.priceService.getPrice()).priceMap;
 
+    // Get signer account
+    let signer = address;
+    const signerSubstrateProxyAddress = transactionInput.signerSubstrateProxyAddress;
+
     if (!transactionInput.skipFeeRecalculation) {
-      validationResponse.estimateFee = await estimateFeeForTransaction(validationResponse, transaction, chainInfo, evmApi, substrateApi, priceMap, feeInfo, nativeTokenInfo, nonNativeTokenPayFeeInfo, transactionInput.isTransferLocalTokenAndPayThatTokenAsFee);
+      validationResponse.estimateFee = await estimateFeeForTransaction(validationResponse, transaction, chainInfo, evmApi, substrateApi, priceMap, feeInfo, nativeTokenInfo, nonNativeTokenPayFeeInfo, transactionInput.isTransferLocalTokenAndPayThatTokenAsFee, signerSubstrateProxyAddress);
     }
 
     const chainInfoMap = this.state.chainService.getChainInfoMap();
 
-    // Check account signing transaction
+    let substrateProxyAccountNativeTokenAvailable: AmountData | undefined;
 
-    checkSigningAccountForTransaction(validationResponse, chainInfoMap);
+    if (signerSubstrateProxyAddress && !isSameAddress(signerSubstrateProxyAddress, address)) {
+      signer = signerSubstrateProxyAddress;
+      substrateProxyAccountNativeTokenAvailable = await this.state.balanceService.getTransferableBalance(signerSubstrateProxyAddress, chain, nativeTokenInfo.slug, extrinsicType);
+    }
+
+    // Check account signing transaction
+    checkSigningAccountForTransaction(validationResponse, chainInfoMap, signer);
 
     const nativeTokenAvailable = await this.state.balanceService.getBalanceByType(address, chain, nativeTokenInfo.slug, transactionInput.balanceType, extrinsicType);
 
     // Check available balance against transaction fee
-    checkBalanceWithTransactionFee(validationResponse, transactionInput, nativeTokenInfo, nativeTokenAvailable);
+    checkBalanceWithTransactionFee(validationResponse, transactionInput, nativeTokenInfo, nativeTokenAvailable, substrateProxyAccountNativeTokenAvailable);
 
     // Warnings Ton address if bounceable and not active
     // if (transaction && isTonTransaction(transaction) && tonApi) {
@@ -810,9 +820,11 @@ export default class TransactionService {
       blockHash: '', // Will be added in next step
       nonce: nonce ?? 0,
       startBlock: startBlock || 0,
-      processId: transaction.step?.processId
+      processId: transaction.step?.processId,
+      substrateProxyAddresses: []
     };
 
+    const substrateProxyHistories: TransactionHistoryItem[] = [];
     const nativeAsset = _getChainNativeTokenBasicInfo(chainInfo);
     const baseNativeAmount = { value: '0', decimals: nativeAsset.decimals, symbol: nativeAsset.symbol };
 
@@ -1164,6 +1176,66 @@ export default class TransactionService {
         break;
       }
 
+      case ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT: {
+        const data = parseTransactionData<ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT>(transaction.data);
+
+        const substrateProxyAddress = data.substrateProxyAddress;
+
+        historyItem.substrateProxyAddresses = [substrateProxyAddress];
+
+        substrateProxyHistories.push({
+          ...historyItem,
+          substrateProxyAddresses: [substrateProxyAddress]
+        });
+
+        try {
+          const substrateProxyAccount = keyring.getPair(substrateProxyAddress);
+
+          if (substrateProxyAccount) {
+            substrateProxyHistories.push({
+              ...historyItem,
+              address: substrateProxyAccount.address,
+              direction: TransactionDirection.RECEIVED,
+              substrateProxyAddresses: [substrateProxyAddress]
+            });
+          }
+        } catch (e) {
+          // skip
+        }
+
+        break;
+      }
+
+      case ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT: {
+        const data = parseTransactionData<ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT>(transaction.data);
+
+        for (const substrateProxyItem of data.selectedSubstrateProxyAccounts || []) {
+          const substrateProxyAddress = substrateProxyItem.substrateProxyAddress;
+
+          substrateProxyHistories.push({
+            ...historyItem,
+            substrateProxyAddresses: [substrateProxyAddress]
+          });
+
+          try {
+            const substrateProxyAccount = keyring.getPair(substrateProxyAddress);
+
+            if (substrateProxyAccount) {
+              substrateProxyHistories.push({
+                ...historyItem,
+                address: substrateProxyAccount.address,
+                direction: TransactionDirection.RECEIVED,
+                substrateProxyAddresses: [substrateProxyAddress]
+              });
+            }
+          } catch (e) {
+            // skip
+          }
+        }
+
+        break;
+      }
+
       case ExtrinsicType.UNKNOWN:
         break;
     }
@@ -1197,7 +1269,7 @@ export default class TransactionService {
       console.warn(e);
     }
 
-    return [historyItem];
+    return [historyItem, ...substrateProxyHistories];
   }
 
   private onSigned ({ id }: TransactionEventResponse) {
@@ -1788,7 +1860,7 @@ export default class TransactionService {
     return emitter;
   }
 
-  private signAndSendSubstrateTransaction ({ address, chain, feeCustom, id, signAfterCreate, step, tokenPayFeeSlug, transaction, url }: SWTransaction): TransactionEmitter {
+  private async signAndSendSubstrateTransaction ({ address, chain, feeCustom, id, signAfterCreate, signerSubstrateProxyAddress, step, tokenPayFeeSlug, transaction, url }: SWTransaction): Promise<TransactionEmitter> {
     const tip = (feeCustom as SubstrateTipInfo)?.tip || '0';
     const feeAssetId = tokenPayFeeSlug && !_isNativeTokenBySlug(tokenPayFeeSlug) && _SUPPORT_TOKEN_PAY_FEE_GROUP.assetHub.includes(chain) ? this.state.chainService.getAssetBySlug(tokenPayFeeSlug).metadata?.multilocation as Record<string, any> : undefined;
 
@@ -1801,14 +1873,26 @@ export default class TransactionService {
       processId: step?.processId
     };
 
-    const extrinsic = transaction as SubmittableExtrinsic;
+    let extrinsic = transaction as SubmittableExtrinsic;
+
+    let signer = address;
+
+    if (signerSubstrateProxyAddress && signerSubstrateProxyAddress !== address) {
+      const substrateApi = this.state.chainService.getSubstrateApi(chain);
+
+      await substrateApi.isReady;
+
+      signer = signerSubstrateProxyAddress;
+      extrinsic = substrateApi.api.tx.proxy.proxy(address, null, transaction as SubmittableExtrinsic);
+    }
+
     // const registry = extrinsic.registry;
     // const signedExtensions = registry.signedExtensions;
 
     const signerOption: Partial<SignerOptions> = {
       signer: {
         signPayload: async (payload: SignerPayloadJSON) => {
-          const { signature, signedTransaction } = await this.state.requestService.signInternalTransaction(id, address, url || EXTENSION_REQUEST_URL, payload, signAfterCreate);
+          const { signature, signedTransaction } = await this.state.requestService.signInternalTransaction(id, signer, url || EXTENSION_REQUEST_URL, payload, signAfterCreate);
 
           return {
             id: (new Date()).getTime(),
@@ -1831,7 +1915,7 @@ export default class TransactionService {
     //   }
     // }
 
-    extrinsic.signAsync(address, signerOption).then(async (rs) => {
+    extrinsic.signAsync(signer, signerOption).then(async (rs) => {
       // Emit signed event
       emitter.emit('signed', eventData);