+++ packages/extension-base/src/utils/account/transform.ts
@@ -133,7 +133,9 @@ export const getAccountActions = (signMode: AccountSignMode, networkType: Accoun
   // QR
   if (signMode === AccountSignMode.PASSWORD) {
     if (networkType === AccountChainType.ETHEREUM || networkType === AccountChainType.SUBSTRATE) {
-      result.push(AccountActions.EXPORT_QR);
+      if (type !== 'ed25519-tw') { // todo: recheck if can export this account type
+        result.push(AccountActions.EXPORT_QR);
+      }
     }
   }
 
@@ -244,6 +246,11 @@ const OPEN_GOV_ACTIONS: ExtrinsicType[] = [
   ExtrinsicType.GOV_UNLOCK_VOTE
 ];
 
+const SUBSTRATE_PROXY_ACTION: ExtrinsicType[] = [
+  ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT,
+  ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT
+];
+
 const OTHER_ACTIONS: ExtrinsicType[] = [
   ExtrinsicType.TRANSFER_XCM,
   ExtrinsicType.SEND_NFT,
@@ -266,8 +273,9 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...EARN_VMANTA_ACTIONS,
           ...CLAIM_AVAIL_BRIDGE,
           ...OPEN_GOV_ACTIONS,
-          ...OTHER_ACTIONS,
-          ExtrinsicType.CHANGE_BITTENSOR_ROOT_CLAIM_TYPE
+          ExtrinsicType.CHANGE_BITTENSOR_ROOT_CLAIM_TYPE,
+          ...SUBSTRATE_PROXY_ACTION,
+          ...OTHER_ACTIONS
         ];
       case AccountChainType.ETHEREUM:
         return [
@@ -277,6 +285,7 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...EARN_STDOT_ACTIONS,
           ...OTHER_ACTIONS,
           ...CLAIM_AVAIL_BRIDGE,
+          ...SUBSTRATE_PROXY_ACTION,
           ...EVM_ACTIONS
         ];
       case AccountChainType.TON:
@@ -308,7 +317,9 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...CLAIM_AVAIL_BRIDGE,
           ...OPEN_GOV_ACTIONS,
           ...OTHER_ACTIONS,
-          ExtrinsicType.CHANGE_BITTENSOR_ROOT_CLAIM_TYPE
+          ExtrinsicType.CHANGE_BITTENSOR_ROOT_CLAIM_TYPE,
+          ...SUBSTRATE_PROXY_ACTION,
+          ...OTHER_ACTIONS
         ];
       case AccountChainType.ETHEREUM:
         return [
@@ -346,8 +357,9 @@ export const getAccountTransactionActions = (signMode: AccountSignMode, networkT
           ...EARN_SDOT_ACTIONS,
           // ...EARN_QDOT_ACTIONS,
           ...OPEN_GOV_ACTIONS,
-          ...OTHER_ACTIONS,
-          ExtrinsicType.CHANGE_BITTENSOR_ROOT_CLAIM_TYPE
+          ExtrinsicType.CHANGE_BITTENSOR_ROOT_CLAIM_TYPE,
+          ...SUBSTRATE_PROXY_ACTION,
+          ...OTHER_ACTIONS
         ];
       case AccountChainType.ETHEREUM:
         return [
@@ -423,6 +435,7 @@ export const getAccountTokenTypes = (type: KeypairType): _AssetType[] => {
       return [_AssetType.NATIVE, _AssetType.LOCAL, _AssetType.ERC20, _AssetType.ERC721];
     case 'sr25519':
     case 'ed25519':
+    case 'ed25519-tw':
     case 'ecdsa':
       return [_AssetType.NATIVE, _AssetType.LOCAL, _AssetType.PSP22, _AssetType.PSP34, _AssetType.GRC20, _AssetType.ERC721, _AssetType.VFT];
     case 'ton':