+++ packages/extension-koni-ui/src/Popup/Confirmations/variants/Transaction/variants/SendNft.tsx
@@ -50,6 +50,13 @@ const Component: React.FC<Props> = (props: Props) => {
             </MetaInfo.Default>
           )
         }
+        {!!transaction.signerSubstrateProxyAddress && transaction.signerSubstrateProxyAddress !== transaction.address &&
+          <MetaInfo.Account
+            address={transaction.signerSubstrateProxyAddress}
+            chainSlug={transaction.chain}
+            label={t('ui.TRANSACTION.Confirmations.SendNft.signWith')}
+          />
+        }
         <MetaInfo.Number
           decimals={decimals}
           label={t('ui.TRANSACTION.Confirmations.SendNft.estimatedFee')}