+++ packages/extension-koni-ui/src/Popup/Confirmations/variants/Transaction/variants/index.tsx
@@ -22,6 +22,8 @@ export { default as GovVoteTransactionConfirmation } from './GovVote';
 export { default as GovUnvoteTransactionConfirmation } from './GovUnvote';
 export { default as GovUnlockTransactionConfirmation } from './GovUnlock';
 export { default as ChangeBittensorRootClaimRewardTypeTransactionConfirmation } from './ChangeBittensorRootClaimRewardType';
+export { default as AddSubstrateProxyAccountTransactionConfirmation } from './AddSubstrateProxyAccount';
+export { default as RemoveSubstrateProxyAccountTransactionConfirmation } from './RemoveSubstrateProxyAccount';
 
 export * from './TransferBlock';
 export * from './Process';