+++ packages/extension-koni-ui/src/Popup/Home/History/index.tsx
@@ -11,9 +11,9 @@ import { DataContext } from '@subwallet/extension-koni-ui/contexts/DataContext';
 import { useFilterModal, useHistorySelection, useSelector, useSetCurrentPage } from '@subwallet/extension-koni-ui/hooks';
 import { cancelSubscription, subscribeTransactionHistory } from '@subwallet/extension-koni-ui/messaging';
 import { SessionStorage, ThemeProps, TransactionHistoryDisplayData, TransactionHistoryDisplayItem } from '@subwallet/extension-koni-ui/types';
-import { customFormatDate, formatHistoryDate, isTypeGov, isTypeStaking, isTypeTransfer } from '@subwallet/extension-koni-ui/utils';
+import { customFormatDate, formatHistoryDate, isTypeGov, isTypeManageSubstrateProxy, isTypeStaking, isTypeTransfer } from '@subwallet/extension-koni-ui/utils';
 import { ButtonProps, Icon, ModalContext, SwIconProps, SwList, SwSubHeader } from '@subwallet/react-ui';
-import { Aperture, ArrowDownLeft, ArrowsLeftRight, ArrowUpRight, Clock, ClockCounterClockwise, Database, FadersHorizontal, NewspaperClipping, Pencil, Rocket, Spinner } from 'phosphor-react';
+import { Aperture, ArrowDownLeft, ArrowsLeftRight, ArrowUpRight, Clock, ClockCounterClockwise, Database, FadersHorizontal, NewspaperClipping, Pencil, Rocket, Spinner, TreeStructure } from 'phosphor-react';
 import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
@@ -35,7 +35,8 @@ const IconMap: Record<string, SwIconProps['phosphorIcon']> = {
   timeout: ClockCounterClockwise,
   swap: ArrowsLeftRight,
   nominate: Pencil,
-  gov: NewspaperClipping
+  gov: NewspaperClipping,
+  substrateProxy: TreeStructure
 };
 
 function getIcon (item: TransactionHistoryItem): SwIconProps['phosphorIcon'] {
@@ -73,6 +74,10 @@ function getIcon (item: TransactionHistoryItem): SwIconProps['phosphorIcon'] {
 
   if (isTypeGov(item.type)) {
     return IconMap.gov;
+  }
+
+  if (isTypeManageSubstrateProxy(item.type)) {
+    return IconMap.substrateProxy;
   }
 
   return IconMap.default;
@@ -339,6 +344,8 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
     [ExtrinsicType.GOV_VOTE]: t('ui.HISTORY.screen.History.vote'),
     [ExtrinsicType.GOV_UNVOTE]: t('ui.HISTORY.screen.History.unvote'),
     [ExtrinsicType.GOV_UNLOCK_VOTE]: t('ui.HISTORY.screen.History.unlockVotes'),
+    [ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT]: t('ui.HISTORY.screen.History.addSubstrateProxy'),
+    [ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT]: t('ui.HISTORY.screen.History.removeSubstrateProxy'),
     [ExtrinsicType.UNKNOWN]: t('ui.HISTORY.screen.History.unknown')
   }), [t]);
 
@@ -389,6 +396,8 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
     [ExtrinsicType.GOV_VOTE]: t('ui.HISTORY.screen.History.voteTransaction'),
     [ExtrinsicType.GOV_UNVOTE]: t('ui.HISTORY.screen.History.unvoteTransaction'),
     [ExtrinsicType.GOV_UNLOCK_VOTE]: t('ui.HISTORY.screen.History.unlockVotesTransaction'),
+    [ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT]: t('ui.HISTORY.screen.History.addSubstrateProxyTransaction'),
+    [ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT]: t('ui.HISTORY.screen.History.removeSubstrateProxyTransaction'),
     [ExtrinsicType.UNKNOWN]: t('ui.HISTORY.screen.History.unknownTransaction')
   }), [t]);
 
@@ -397,9 +406,13 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
     const finalHistoryMap: Record<string, TransactionHistoryDisplayItem> = {};
 
     rawHistoryList.forEach((item: TransactionHistoryItem) => {
-      // Format display name for account by address
       const fromName = accountMap[quickFormatAddressToCompare(item.from) || ''];
-      const toName = accountMap[quickFormatAddressToCompare(item.to) || ''];
+      let toName = accountMap[quickFormatAddressToCompare(item.to) || ''];
+
+      if ((item.type === ExtrinsicType.ADD_SUBSTRATE_PROXY_ACCOUNT || item.type === ExtrinsicType.REMOVE_SUBSTRATE_PROXY_ACCOUNT) && item.substrateProxyAddresses?.length) {
+        toName = accountMap[quickFormatAddressToCompare(item.address) || ''];
+      }
+
       const key = getHistoryItemKey(item);
       const displayTime = item.blockTime || item.time;