+++ packages/extension-koni-ui/src/Popup/Transaction/variants/CancelUnstake.tsx
@@ -46,7 +46,7 @@ const Component = () => {
   const { t } = useTranslation();
   const navigate = useNavigate();
 
-  const { defaultData, persistData } = useTransactionContext<CancelUnStakeParams>();
+  const { defaultData, persistData, selectSubstrateProxyAccountsToSign } = useTransactionContext<CancelUnStakeParams>();
   const { slug } = defaultData;
 
   const [form] = Form.useForm<CancelUnStakeParams>();
@@ -66,7 +66,6 @@ const Component = () => {
 
   const { list: allPositionInfos } = useYieldPositionDetail(slug);
   const { compound: positionInfo } = useYieldPositionDetail(slug, fromValue);
-
   const [isDisable, setIsDisable] = useState(true);
   const [loading, setLoading] = useState(false);
   const [isBalanceReady, setIsBalanceReady] = useState(true);
@@ -108,23 +107,43 @@ const Component = () => {
 
     setLoading(true);
 
-    const { from, slug, unstake: unstakeIndex } = values;
+    const { chain, from, slug, unstake: unstakeIndex } = values;
 
     const selectedUnstaking = positionInfo.unstakings[parseInt(unstakeIndex)];
 
-    setTimeout(() => {
-      yieldSubmitStakingCancelWithdrawal({
+    // send cancel unstake transaction
+    const sendPromise = (signerSubstrateProxyAddress?: string) => {
+      return yieldSubmitStakingCancelWithdrawal({
         address: from,
         slug,
-        selectedUnstaking
-      })
-        .then(onSuccess)
-        .catch(onError)
-        .finally(() => {
-          setLoading(false);
+        selectedUnstaking,
+        signerSubstrateProxyAddress
+      }).then(onSuccess);
+    };
+
+    // wrap proxy selection
+    // for the Liquid Staking feature with multiple steps,
+    // only the root account is allowed to sign transactions, even if a valid proxy account is available to sign on its behalf.
+    const sendPromiseWrapper = async () => {
+      if (poolInfo.type !== YieldPoolType.LIQUID_STAKING) {
+        const substrateProxyAddress = await selectSubstrateProxyAccountsToSign({
+          chain,
+          address: from,
+          type: ExtrinsicType.STAKING_CANCEL_UNSTAKE
         });
+
+        return await sendPromise(substrateProxyAddress);
+      }
+
+      return await sendPromise();
+    };
+
+    // delay for better loading UX
+    setTimeout(() => {
+      sendPromiseWrapper().catch(onError)
+        .finally(() => setLoading(false));
     }, 300);
-  }, [onError, onSuccess, positionInfo]);
+  }, [positionInfo, onSuccess, poolInfo.type, selectSubstrateProxyAccountsToSign, onError]);
 
   const onPreCheck = usePreCheckAction(fromValue);