+++ packages/extension-koni-ui/src/components/Modal/Earning/EarningValidatorSelectedModal/ChangeBittensorValidator.tsx
@@ -4,7 +4,7 @@
 import { TransactionError } from '@subwallet/extension-base/background/errors/TransactionError';
 import { ExtrinsicType, NotificationType, ValidatorInfo } from '@subwallet/extension-base/background/KoniTypes';
 import { isActionFromValidator } from '@subwallet/extension-base/services/earning-service/utils';
-import { NominationInfo, SubmitBittensorChangeValidatorStaking, YieldPoolType } from '@subwallet/extension-base/types';
+import { NominationInfo, YieldPoolType } from '@subwallet/extension-base/types';
 import { MetaInfo } from '@subwallet/extension-koni-ui/components';
 import { BasicInputWrapper } from '@subwallet/extension-koni-ui/components/Field/Base';
 import { WalletModalContext } from '@subwallet/extension-koni-ui/contexts/WalletModalContextProvider';
@@ -61,8 +61,7 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
 
   const { alertModal: { close: closeAlert, open: openAlert } } = useContext(WalletModalContext);
   const { defaultData } = useTransactionContext<ChangeValidatorParams>();
-  const { onError, onSuccess } = useHandleSubmitTransaction();
-
+  const { onError, onSuccess, selectSubstrateProxyAccountsToSign } = useHandleSubmitTransaction();
   const account = findAccountByAddress(accounts, from);
   const [form] = Form.useForm<ChangeValidatorParams>();
   const originValidator = useWatchTransaction('originValidator', form, defaultData);
@@ -284,12 +283,21 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
       const send = (amount: string): void => {
         setSubmitLoading(true);
 
-        const submitData: SubmitBittensorChangeValidatorStaking = {
-          ...baseData,
-          amount
+        // send change earning validator transaction
+        const sendPromise = (signerSubstrateProxyAddress?: string) => {
+          return changeEarningValidator({
+            ...baseData,
+            amount,
+            signerSubstrateProxyAddress
+          });
         };
 
-        (changeEarningValidator(submitData))
+        // select proxy account if available
+        selectSubstrateProxyAccountsToSign({
+          chain,
+          address: from,
+          type: ExtrinsicType.CHANGE_EARNING_VALIDATOR
+        }).then(sendPromise)
           .then(onSuccess)
           .catch((error: TransactionError) => {
             if (error.message.includes('remaining')) {
@@ -324,7 +332,7 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
 
       send(isShowAmountChange ? value : bondedValue);
     },
-    [bondedValue, closeAlert, from, isShowAmountChange, netuid, notifyTooHighAmount, onError, onSuccess, openAlert, poolInfo.slug, poolTargets, stakingFee, symbol, t]
+    [bondedValue, chain, closeAlert, from, isShowAmountChange, netuid, notifyTooHighAmount, onError, onSuccess, openAlert, poolInfo.slug, poolTargets, selectSubstrateProxyAccountsToSign, stakingFee, symbol, t]
   );
   const { onCancelSelectValidator } = useSelectValidators(modalId, chain, maxCount, onChange, isSingleSelect);