+++ packages/extension-koni-ui/src/components/index.ts
@@ -39,3 +39,4 @@ export * from './WalletConnect';
 export * from './Crowdloan';
 export * from './TransactionProcess';
 export * from './Governance';
+export * from './SubstrateProxyAccount';