+++ packages/extension-koni-ui/src/constants/modal.ts
@@ -9,6 +9,7 @@ export const IMPORT_ACCOUNT_MODAL = 'import-account-modal';
 export const ATTACH_ACCOUNT_MODAL = 'attach-account-modal';
 export const NEW_SEED_MODAL = 'new-seed-modal';
 export const IMPORT_SEED_MODAL = 'import-seed-modal';
+export const IMPORT_SEED_TRUST_MODAL = 'import-seed-trust-modal';
 export const REQUEST_CREATE_PASSWORD_MODAL = 'request-create-password-modal';
 export const CONFIRMATION_DETAIL_MODAL = 'confirmation-detail-modal';
 export const CONFIRMATION_QR_MODAL = 'confirmation-qr-modal';
@@ -95,6 +96,11 @@ export const TRANSACTION_STEPS_MODAL = 'transaction-steps-modal';
 /* OpenGov */
 export const GOV_DELEGATION_DETAILS_MODAL = 'gov-delegation-details-modal';
 /* OpenGov */
+
+/* Proxy */
+export const SUBSTRATE_PROXY_ACCOUNT_SELECTOR_MODAL = 'substrateProxyAccount-account-selector-modal';
+export const SUBSTRATE_PROXY_ACCOUNT_LIST_MODAL = 'substrateProxyAccount-account-list-modal';
+/* Proxy */
 
 /* Other */
 export const EARNING_SLIPPAGE_MODAL = 'earning-slippage-modal';