+++ packages/extension-koni-ui/src/types/transaction.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { StakingType } from '@subwallet/extension-base/background/KoniTypes';
+import { SubstrateProxyType } from '@subwallet/extension-base/types';
 
 export interface TransactionFormBaseProps {
   fromAccountProxy: string;
@@ -98,4 +99,12 @@ export interface GovUnlockVoteParams extends TransactionFormBaseProps {
 
 export interface ChangeBittensorRootClaimType extends TransactionFormBaseProps {
   bittensorRootClaimType: string;
+}
+export interface AddSubstrateProxyAccountParams extends TransactionFormBaseProps {
+  substrateProxyAddress: string;
+  substrateProxyType: SubstrateProxyType;
+}
+
+export interface RemoveSubstrateProxyAccountParams extends TransactionFormBaseProps {
+  substrateProxyAddressKeys: string[]; // key format is "substrateProxyType:address"
 }