+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # CHANGELOG
 
+## 1.3.72
+Build date: Jan 14, 2025.
+
+Features & Updates:
+- Integrate Proxy Account Support (#4725)
+- Update chain-list stable v0.2.123 (#4861)
+- Add support for USDC & stEWT (#639)
+- Migrate to ParaSpell V5 (#4908)
+
+
+## 1.3.71
+Build date: Dec 29, 2025.
+
+Features & Updates:
+- Improve token enabling (Round 2) (#4247)
+- Update libs for SubWallet Extensions (#4808)
+- Support Import from Trust Wallet to SubWallet (#4762)
+
+
 ## 1.3.70
 Build date: Dec 11, 2025.